% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblender.R
\name{call}
\alias{call}
\title{Make HTTP request to \href{http://openblender.io}{openblender.io} service}
\usage{
call(action, parameters)
}
\arguments{
\item{action}{Task you're requesting}

\item{parameters}{Request options}
}
\value{
The OpenBlender service response, which depends on the action and parameters provided.
}
\description{
Function made to call OpenBlender API services.
}
\examples{
##CREATE A DATASET
df <- read.csv(file = "/path/to/your/data.csv", header = TRUE, sep = ",")
action <- "API_createDataset"
parameters <- list(
token = "YOUR_TOKEN",
id_user = "YOUR_USER_ID",
name = "dataset name",
descriptipon = "Provide a description here",
visibility = "public",
tags = list("topic", "tag"),
insert_observations = "on",
dataframe = df
)
response <- openblender::call(action, parameters)

##INSERT OBSERVATIONS
df <- read.csv(file = "/path/to/your/data.csv", header = TRUE, sep = ",")
action <- "API_insertObservations"
parameters <- list(
token = "YOUR_TOKEN",
id_user = "YOUR_USER_ID",
id_dataset = "DATASET_ID",
notification = "on",
observations = df
)
response <- openblender::call(action, parameters)

##GET OBSERVATIONS
action <- "API_getObservationsFromDataset"
parameters <- list(
token = "YOUR_TOKEN",
id_user = "YOUR_USER_ID",
id_dataset = "DATASET_ID"
)
response <- openblender::call(action, parameters)
}
