% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comprobarJSONaDF}
\alias{comprobarJSONaDF}
\title{Verify JSON}
\usage{
comprobarJSONaDF(df_json)
}
\arguments{
\item{df_json}{Dataframe to verify}
}
\value{
Dataframe verified or an error message.
}
\description{
Check if a dataframe can be transformed into JSON with no errors. This function is not accessible for users.
}
\keyword{internal}
