% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dateToUnix}
\alias{dateToUnix}
\title{Date string to Unix timestamp.}
\usage{
dateToUnix(date_time_str, date_format, timezone = "GMT")
}
\arguments{
\item{date_time_str}{String with a date in any format. For example: ‘6.2.1986 22:11:00'. This can also be a datetime string vector or the datetime column in your anchor.}

\item{date_format}{The format of the date string.}

\item{timezone}{String of timezone, Default GMT. To list possible timezones execute OlsonNames()}
}
\value{
A double unix timestamp or a vector of those doubles.
}
\description{
Returns a Unix timestamp with a date or datetime string. It also accepts a list or vector of strings and returns a converted list of Unix timestamps. Consult documentation \href{https://www.openblender.io/#/api_documentation}{here}.
}
\seealso{
To see more details go to \href{http://openblender.io}{openblender.io}
}
