% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Request to the API, depending on the action provided}
\usage{
get_observations(json_parametros, url)
}
\arguments{
\item{json_parametros}{Request parameters}

\item{url}{Url selected}
}
\value{
List of observations obtained with \link{dameRespuestaLlamado}.
}
\description{
Prepare the data to send it 'OpenBlender' API. This function is not used by users.
}
\keyword{internal}
