% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationBlend.R
\name{locationBlend}
\alias{locationBlend}
\title{Location Blend.}
\usage{
locationBlend(
  token,
  anchor_lat,
  anchor_lon,
  blend_source,
  blend_type = "closest_observation",
  agg_output = "count",
  n = 3,
  r = 1000,
  data_format = "dataframe",
  consumption_confirmation = "off",
  oblender = FALSE
)
}
\arguments{
\item{token}{Obtained in \href{https://www.openblender.io/#/user/account?section=account}{openblender.io}}

\item{anchor_lat}{Array of ‘latitude’ values to blend data by.}

\item{anchor_lon}{Array of ‘longitude’ values to blend data by.}

\item{blend_source}{JSON object of Data To Blend configuration obtained from a data source in OpenBlender.io}

\item{blend_type}{(Only if blend_type = 'agg_in_intervals'). Select chronological direction of observation or interval to aggregate from 'blend source' to 'anchor':’n_closest_agg’: Aggregate closest n. ‘in_radio_agg’: Aggregate in radio r (measured in km). Default: 'in_radio_agg'.}

\item{agg_output}{'sum', 'avg': For numerical features. Default: 'count'}

\item{n}{(Only if ‘n_closest_agg’ is selected) Number of closest observations to blend and aggregate.}

\item{r}{(Only if ‘in_radio_agg’ is selected) Radio in kilometers to blend and aggregate data points.}

\item{data_format}{Default "Dataframe"}

\item{consumption_confirmation}{"on" or "off", default "on"}

\item{oblender}{Omit this parameter}
}
\value{
A Blended dataset intersected with your anchor in location, success/error message when you insert observations or the list of observations requested.
}
\description{
Get Dataframe with DataToBlend ordered time by latitude, sending two lists or Series of Anchor latitude and longitude. Consult documentation \href{https://www.openblender.io/#/api_documentation}{here}.
}
\seealso{
To see more details go to \href{http://openblender.io}{openblender.io}
}
