% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationBlend.R
\name{searchLocationBlends}
\alias{searchLocationBlends}
\title{Search Location Blends.}
\usage{
searchLocationBlends(
  token,
  anchor_lat,
  anchor_lon,
  search_text,
  oblender = FALSE
)
}
\arguments{
\item{token}{Obtained in \href{https://www.openblender.io/#/user/account?section=account}{openblender.io}}

\item{anchor_lat}{Latitude array of your anchor.}

\item{anchor_lon}{Longitude array of your anchor.}

\item{search_text}{Text provided to find Time Blends.}

\item{oblender}{Omit this parameter.}
}
\value{
A list of datasets with ids, features, and more info.
}
\description{
Search Blends that intersect your anchor in location providing a string. Consult documentation \href{https://www.openblender.io/#/api_documentation}{here}.
}
\examples{
\dontrun{
results <- openblender::searchTimeBlends(token="YOUR_TOKEN",
                                        anchor_lat=df_anchor$latitude,
                                        anchor_lon=df_anchor$longitude,
                                        search_text="coronavirus")
}
}
\seealso{
To see more details go to \href{http://openblender.io}{openblender.io}
}
