% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeBlend.R
\name{searchTimeBlends}
\alias{searchTimeBlends}
\title{Search Time Blends.}
\usage{
searchTimeBlends(token, anchor_ts, search_text, oblender = FALSE)
}
\arguments{
\item{token}{Obtained in \href{https://www.openblender.io/#/user/account?section=account}{openblender.io}}

\item{anchor_ts}{Dataframe with a timestamp column in unix timestamp format.}

\item{search_text}{Text provided to find Time Blends.}

\item{oblender}{Omit this parameter.}
}
\value{
A list of datasets with ids, features, and more info.
}
\description{
Search Blends that intersect your anchor in time providing a string. Consult documentation \href{https://www.openblender.io/#/api_documentation}{here}.
}
\examples{
\dontrun{
results <- openblender::searchTimeBlends(token="YOUR_TOKEN",
                                        anchor_ts=df_anchor$timestamp,
                                        search_text="coronavirus")
}
}
\seealso{
To see more details go to \href{http://openblender.io}{openblender.io}
}
