% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeBlend.R
\name{timeBlend}
\alias{timeBlend}
\title{Time Blend.}
\usage{
timeBlend(
  token,
  anchor_ts,
  blend_source,
  blend_type = "closest_observation",
  direction = "time_prior",
  interval_output = "count",
  ts_restriction = NULL,
  oblender = FALSE,
  interval_size = 3600,
  consumption_confirmation = "off",
  missing_values = "raw",
  data_format = "dataframe"
)
}
\arguments{
\item{token}{Obtained in \href{https://www.openblender.io/#/user/account?section=account}{openblender.io}}

\item{anchor_ts}{Array of Unix timestamp values to blend data to.}

\item{blend_source}{JSON object of Data To Blend configuration obtained from a data source in OpenBlender.io}

\item{blend_type}{Type of blend to gather data: 'closest_observation', 'agg_in_intervals', (Default: closest_observation).}

\item{direction}{(Only if blend_type = 'agg_in_intervals'). Select chronological direction of observation or interval to aggregate from 'blend source' to 'anchor': 'time_prior': Gather data from blend source prior or at the same time as each observation from anchor. 'time_prior_strict': Gather data from blend source strictly prior to each observation from anchor. 'time_post': Gather data subsequent to or at the same time than each observation from anchor. 'time_post_strict': Gather data strictly subsequent to each observation from anchor. 'time_prior_and_post': Gather data prior and subsequent to each observation from anchor. Default (time_prior)}

\item{interval_output}{(Only if blend_type = 'agg_in_range'), options: 'count' (any type of feature). 'sum' 'avg' (for numerical features). 'text', 'text_list' (for text features). (Default = count)}

\item{ts_restriction}{None}

\item{oblender}{Omit this parameter}

\item{interval_size}{(Only if blend_type = 'agg_in_intervals'). Time interval size in seconds to be used with the selected agg_interval_order.}

\item{consumption_confirmation}{"on" or "off", default "on"}

\item{missing_values}{Treatment of missing values. 'raw': Ignore missing values (return NaN). 'impute': Impute missing values. (Default 'raw')}

\item{data_format}{Default "Dataframe"}
}
\value{
A Blended dataset intersected with your anchor in time, success/error message when you insert observations or the list of observations requested.
}
\description{
Get Dataframe with DataToBlend ordered time by sending a list or Series of Anchor timestamps. Consult documentation \href{https://www.openblender.io/#/api_documentation}{here}.
}
\examples{
\dontrun{
#Blend a dataset
#You provide df_anchor data
blend_source <- list(
id_dataset="5f063dda198b5854284dae58",
feature="price",
dataset_filter=c(list(low=list("$gt"=300)))
)

df_blend <- openblender::timeBlend(token = "YOUR_TOKEN",
                                  anchor_ts = df_anchor$timestamp,
                                  blend_type = "agg_in_intervals",
                                  direction = "time_prior",
                                  interval_output = "avg",
                                  interval_size = 60 * 60 * 24 * 1,
                                  blend_source = blend_source,
                                  missing_values = "impute")
}
}
\seealso{
To see more details go to \href{http://openblender.io}{openblender.io}
}
