% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unixToDate}
\alias{unixToDate}
\title{Unix timestamp to Date string.}
\usage{
unixToDate(
  unix_timestamp,
  date_format = "\%d-\%m-\%Y \%H:\%M:\%S",
  timezone = "GMT"
)
}
\arguments{
\item{unix_timestamp}{Unix timestamp as double. For example : 508133460 This can also be a vector of those timestamps or the timestamp column in your anchor.}

\item{date_format}{The format of the date string.}

\item{timezone}{String of timezone, Default GMT. To list possible timezones execute OlsonNames()}
}
\value{
A double unix timestamp or a vector of those doubles.
}
\description{
Returns a Unix timestamp with a date or datetime string. It also accepts a list or vector of strings and returns a converted list of Unix timestamps. Consult documentation \href{https://www.openblender.io/#/api_documentation}{here}.
}
\examples{
\dontrun{
df_anchor$timestamp <- openblender::dateToUnix(date_time_str=1602160634,
                                              date_format="\%d.\%m.\%Y \%H:\%M:\%S",
                                              timezone="US/Eastern")
}
}
\seealso{
To see more details go to \href{http://openblender.io}{openblender.io}
}
